requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
py-libp2p>=0.1.5
torch>=1.9.0
pydantic>=1.8.0
cryptography>=3.4.0
aiohttp>=3.8.0
python-multipart>=0.0.5
pytest>=6.2.0
docker>=5.0.0
solana>=0.20.0
psutil>=5.9.0
